Update update_discord.sh
The logic is like this now:
- try to get a new version
- if it's available - unpack and launch it
- if it's not - launch the old one
- if the old one complains about needing an update - patch the `skipped_versions.txt` with the `needed version + 1` value and relaunch the script

The skipped versions file needed as Discord tends to upload broken versions and later rolling them back and forth several times. So instead we block the next version and only update if Discord complains about it. 

Probably the whole download first logic is not needed at this point, we can just react to stdout. IDC, TBH. Maybe'll fix later.